refactor: [route] HubRoute CRUD logic refactoring
