+++ route/src/main/java/com/sparta/route/domain/controller/hubRoute/HubRouteController.java
@@ -1,7 +1,9 @@
-package com.sparta.route.domain.controller;
+package com.sparta.route.domain.controller.hubRoute;
 
-import com.sparta.route.domain.dto.HubRouteDto;
-import com.sparta.route.domain.service.HubRouteService;
+import com.sparta.route.domain.dto.hubRoute.HubRouteDto;
+import com.sparta.route.domain.dto.hubRoute.HubRouteRequestDto;
+import com.sparta.route.domain.service.hubRoute.HubRouteService;
+import jakarta.validation.Valid;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -18,42 +20,43 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("api/v1/routes/hub")
+@RequestMapping("api/v1/hubs/routes")
 public class HubRouteController {
 
     private final HubRouteService hubRouteService;
 
     @PostMapping
-    public ResponseEntity<HubRouteDto> createHubRoute(@RequestBody HubRouteDto hubRouteDto) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(hubRouteService.createHubRoute(hubRouteDto));
+    public ResponseEntity<HubRouteDto> createHubRoute(@Valid @RequestBody HubRouteRequestDto hubRouteRequestDto) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(hubRouteService.createHubRoute(hubRouteRequestDto.toDTO()));
     }
 
-    @PutMapping("/{originHubId}/{destinationHubId}")
+    @PutMapping("/{originHubId}/-/{destinationHubId}")
     public ResponseEntity<HubRouteDto> updateHubRoute(
             @PathVariable UUID originHubId,
             @PathVariable UUID destinationHubId,
-            @RequestBody HubRouteDto hubRouteDto) {
-        return ResponseEntity.ok(hubRouteService.updateHubRoute(originHubId, destinationHubId, hubRouteDto));
+            @Valid @RequestBody HubRouteRequestDto hubRouteRequestDto) {
+        return ResponseEntity.ok(hubRouteService.updateHubRoute(originHubId, destinationHubId,
+                hubRouteRequestDto.toDTO()));
     }
 
-    @DeleteMapping("/{originHubId}/{destinationHubId}")
+    @DeleteMapping("/{originHubId}/-/{destinationHubId}")
     public ResponseEntity<Void> deleteHubRoute(@PathVariable UUID originHubId, @PathVariable UUID destinationHubId) {
         hubRouteService.deleteHubRoute(originHubId, destinationHubId);
         return ResponseEntity.noContent().build();
     }
 
-    @GetMapping("/{originHubId}/{destinationHubId}")
+    @GetMapping("/{originHubId}/-/{destinationHubId}")
     public ResponseEntity<HubRouteDto> getHubRoute(@PathVariable UUID originHubId, @PathVariable UUID destinationHubId) {
         return ResponseEntity.ok(hubRouteService.getHubRoute(originHubId, destinationHubId));
-    }
-
-    @GetMapping
-    public ResponseEntity<List<HubRouteDto>> getAllHubRoutes() {
-        return ResponseEntity.ok(hubRouteService.getAllHubRoutes());
     }
 
     @GetMapping("/origin/{originHubId}")
     public ResponseEntity<List<HubRouteDto>> getHubRoutesByOrigin(@PathVariable UUID originHubId) {
         return ResponseEntity.ok(hubRouteService.getHubRoutesByOrigin(originHubId));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<HubRouteDto>> getAllHubRoutes() {
+        return ResponseEntity.ok(hubRouteService.getAllHubRoutes());
     }
 }
